chore: Move TestSetGradesIfApproved to submission_test.go
